+++ packages/components-js/tests/e2e/specs/select.e2e.ts
@@ -28,7 +28,11 @@ const getButton = (page: Page) => page.locator('p-select button').first();
 const getButtonImage = (page: Page) => page.locator('p-select button img').first().getAttribute('src');
 const getDropdown = (page: Page) => page.locator('p-select [popover]');
 const getDropdownDisplay = async (page: Page): Promise<string> => await getElementStyle(getDropdown(page), 'display');
+const getFilter = (page: Page) => page.locator('p-select p-input-search');
+const getFilterInput = (page: Page) => page.locator('p-select p-input-search input');
 const getSelectOption = (page: Page, n: number) => page.locator(`p-select p-select-option:nth-child(${n})`);
+const getSelectOptions = (page: Page) => page.locator('p-select p-select-option');
+const getSelectOptgroups = (page: Page) => page.locator('p-select p-optgroup');
 const getSelectedSelectOptionProperty = async <K extends keyof SelectOption>(
   page: Page,
   property: K
@@ -475,6 +479,19 @@ test.describe('focus', () => {
 
     expect((await getEventSummary(buttonElement, 'focus')).counter).toBe(1);
     await expect(buttonElement).toHaveCSS('outline-color', 'rgb(26, 68, 234)');
+  });
+  test('should focus filter input after opening when filter is enabled', async ({ page }) => {
+    await initSelect(page, { props: { name: 'Some name', filter: true } });
+    const filter = getFilter(page);
+
+    await expect(filter).toBeHidden();
+    await expect(filter).not.toBeFocused();
+
+    await page.keyboard.press('Tab');
+    await page.keyboard.press('Space');
+
+    await expect(filter).toBeVisible();
+    await expect(filter).toBeFocused();
   });
 });
 
@@ -1173,6 +1190,511 @@ test.describe('selection', () => {
   });
 });
 
+test.describe('filter', () => {
+  test.describe('input', () => {
+    test('should show matching options when typing into filter', async ({ page }) => {
+      await initSelect(page, { props: { name: 'Some name', filter: true } });
+      const buttonElement = getButton(page);
+      const filterElement = getFilter(page);
+      const filterInputElement = getFilterInput(page);
+      const options = getSelectOptions(page);
+      const dropdown = getDropdown(page);
+
+      await buttonElement.click();
+
+      await expect(dropdown).toBeVisible();
+      await expect(filterElement).toBeFocused();
+      await expect(filterInputElement).toHaveValue('');
+
+      await expect(options).toHaveCount(3);
+      await expect(options.nth(0)).toBeVisible();
+      await expect(options.nth(1)).toBeVisible();
+      await expect(options.nth(2)).toBeVisible();
+
+      await filterInputElement.fill('b');
+
+      await expect(options.nth(0)).toBeHidden();
+      await expect(options.nth(1)).toBeVisible();
+      await expect(options.nth(1)).toHaveText('b');
+      await expect(options.nth(2)).toBeHidden();
+
+      await filterInputElement.fill('');
+
+      await expect(options.nth(0)).toBeVisible();
+      await expect(options.nth(1)).toBeVisible();
+      await expect(options.nth(2)).toBeVisible();
+    });
+
+    test('should reset filter when pressing clear button on filter input', async ({ page }) => {
+      await initSelect(page, { props: { name: 'Some name', filter: true } });
+      const buttonElement = getButton(page);
+      const filterElement = getFilter(page);
+      const filterInputElement = getFilterInput(page);
+      const options = getSelectOptions(page);
+      const dropdown = getDropdown(page);
+
+      await buttonElement.click();
+
+      await expect(dropdown).toBeVisible();
+      await expect(filterElement).toBeFocused();
+      await expect(filterInputElement).toHaveValue('');
+      await filterInputElement.fill('b');
+
+      await expect(options.nth(0)).toBeHidden();
+      await expect(options.nth(1)).toBeVisible();
+      await expect(options.nth(1)).toHaveText('b');
+      await expect(options.nth(2)).toBeHidden();
+
+      await filterElement.locator('p-button-pure').click();
+      await expect(filterInputElement).toHaveValue('');
+
+      await expect(options.nth(0)).toBeVisible();
+      await expect(options.nth(1)).toBeVisible();
+      await expect(options.nth(2)).toBeVisible();
+    });
+
+    skipInBrowsers(['webkit'], () => {
+      test('should reset filter value and show all options again after filtering and selecting an option', async ({
+        page,
+      }) => {
+        await initSelect(page, { props: { name: 'Some name', filter: true } });
+        const host = getHost(page);
+        const buttonElement = getButton(page);
+        const filterElement = getFilter(page);
+        const filterInputElement = getFilterInput(page);
+        const options = getSelectOptions(page);
+        const dropdown = getDropdown(page);
+
+        await buttonElement.click();
+
+        await expect(dropdown).toBeVisible();
+        await expect(filterElement).toBeFocused();
+        await expect(filterInputElement).toHaveValue('');
+        await filterInputElement.fill('b');
+
+        await expect(options.nth(0)).toBeHidden();
+        await expect(options.nth(1)).toBeVisible();
+        await expect(options.nth(1)).toHaveText('b');
+        await expect(options.nth(2)).toBeHidden();
+
+        await page.keyboard.press('ArrowDown');
+        await expect(options.nth(1)).toHaveJSProperty('highlighted', true);
+        await page.keyboard.press('Enter');
+
+        await expect(dropdown).toBeHidden();
+        await expect(host).toHaveJSProperty('value', 'b');
+        await expect(buttonElement).toBeFocused();
+
+        await page.keyboard.press('Space');
+        await expect(filterElement).toBeFocused();
+        await expect(filterInputElement).toHaveValue('');
+
+        await expect(options.nth(0)).toBeVisible();
+        await expect(options.nth(1)).toBeVisible();
+        await expect(options.nth(2)).toBeVisible();
+      });
+    });
+
+    skipInBrowsers(['webkit', 'firefox'], () => {
+      test('should reset filter value and show all options again after closing and reopening again', async ({
+        page,
+      }) => {
+        await initSelect(page, { props: { name: 'Some name', filter: true } });
+        const host = getHost(page);
+        const buttonElement = getButton(page);
+        const filterElement = getFilter(page);
+        const filterInputElement = getFilterInput(page);
+        const options = getSelectOptions(page);
+        const dropdown = getDropdown(page);
+
+        await buttonElement.click();
+
+        await expect(dropdown).toBeVisible();
+        await expect(filterElement).toBeFocused();
+        await expect(filterInputElement).toHaveValue('');
+        await filterInputElement.fill('b');
+
+        await page.keyboard.press('Escape');
+        await expect(dropdown).toBeHidden();
+        await expect(host).toHaveJSProperty('value', undefined);
+        await expect(buttonElement).toBeFocused();
+
+        await waitForStencilLifecycle(page);
+        await buttonElement.click();
+        await expect(filterElement).toBeFocused();
+        await expect(filterInputElement).toHaveValue('');
+
+        await expect(options.nth(0)).toBeVisible();
+        await expect(options.nth(1)).toBeVisible();
+        await expect(options.nth(2)).toBeVisible();
+      });
+    });
+
+    skipInBrowsers(['webkit'], () => {
+      test('should show indicator when no results are found', async ({ page }) => {
+        await initSelect(page, { props: { name: 'Some name', filter: true } });
+        const buttonElement = getButton(page);
+        const filterElement = getFilter(page);
+        const filterInputElement = getFilterInput(page);
+        const options = getSelectOptions(page);
+        const dropdown = getDropdown(page);
+
+        await buttonElement.click();
+
+        await expect(dropdown).toBeVisible();
+        await expect(filterElement).toBeFocused();
+        await expect(filterInputElement).toHaveValue('');
+        await filterInputElement.fill('no results');
+
+        for (const option of await options.all()) {
+          await expect(option).toBeHidden();
+        }
+
+        const noResults = page.getByRole('option', { name: 'No results found' });
+        await expect(noResults).toBeVisible();
+
+        await page.keyboard.press('Escape');
+        await expect(dropdown).toBeHidden();
+        await expect(buttonElement).toBeFocused();
+
+        await waitForStencilLifecycle(page);
+        await buttonElement.click();
+        await expect(filterElement).toBeFocused();
+        await expect(filterInputElement).toHaveValue('');
+
+        await expect(noResults).toBeHidden();
+        await expect(options.nth(0)).toBeVisible();
+        await expect(options.nth(1)).toBeVisible();
+        await expect(options.nth(2)).toBeVisible();
+      });
+    });
+  });
+
+  test.describe('with optgroups', () => {
+    test('should only show optgroups of matching options when filtering', async ({ page }) => {
+      await initSelect(page, {
+        props: {
+          name: 'Some name',
+          filter: true,
+        },
+        options: {
+          includeOptgroups: true,
+          values: [[{ value: '1a' }], [{ value: '2a' }]],
+        },
+      });
+      const buttonElement = getButton(page);
+      const filterElement = getFilter(page);
+      const filterInputElement = getFilterInput(page);
+      const options = getSelectOptions(page);
+      const optgroups = getSelectOptgroups(page);
+      const dropdown = getDropdown(page);
+
+      await buttonElement.click();
+
+      await expect(dropdown).toBeVisible();
+      await expect(filterElement).toBeFocused();
+      await expect(filterInputElement).toHaveValue('');
+
+      await expect(optgroups).toHaveCount(2);
+      await expect(options).toHaveCount(2);
+      await expect(optgroups.nth(0)).toBeVisible();
+      await expect(options.nth(0)).toBeVisible();
+      await expect(optgroups.nth(1)).toBeVisible();
+      await expect(options.nth(1)).toBeVisible();
+
+      await filterInputElement.fill('2');
+
+      await expect(optgroups).toHaveCount(2);
+      await expect(options).toHaveCount(2);
+      await expect(optgroups.nth(0)).toBeHidden();
+      await expect(options.nth(0)).toBeHidden();
+      await expect(optgroups.nth(1)).toBeVisible();
+      await expect(options.nth(1)).toBeVisible();
+
+      await filterInputElement.fill('a');
+
+      await expect(optgroups).toHaveCount(2);
+      await expect(options).toHaveCount(2);
+      await expect(optgroups.nth(0)).toBeVisible();
+      await expect(options.nth(0)).toBeVisible();
+      await expect(optgroups.nth(1)).toBeVisible();
+      await expect(options.nth(1)).toBeVisible();
+    });
+
+    test('should reset filter value and show all optgroups and options again after filtering and selecting an option', async ({
+      page,
+    }) => {
+      await initSelect(page, {
+        props: {
+          name: 'Some name',
+          filter: true,
+        },
+        options: {
+          includeOptgroups: true,
+          values: [[{ value: '1a' }], [{ value: '2a' }]],
+        },
+      });
+      const host = getHost(page);
+      const buttonElement = getButton(page);
+      const filterElement = getFilter(page);
+      const filterInputElement = getFilterInput(page);
+      const options = getSelectOptions(page);
+      const optgroups = getSelectOptgroups(page);
+      const dropdown = getDropdown(page);
+
+      await buttonElement.click();
+
+      await expect(dropdown).toBeVisible();
+      await expect(filterElement).toBeFocused();
+      await filterInputElement.fill('2');
+
+      await expect(optgroups).toHaveCount(2);
+      await expect(options).toHaveCount(2);
+      await expect(optgroups.nth(0)).toBeHidden();
+      await expect(options.nth(0)).toBeHidden();
+      await expect(optgroups.nth(1)).toBeVisible();
+      await expect(options.nth(1)).toBeVisible();
+
+      await options.nth(1).click();
+
+      await expect(dropdown).toBeHidden();
+      await expect(host).toHaveJSProperty('value', '2a');
+      await expect(buttonElement).toBeFocused();
+
+      await buttonElement.click();
+
+      await expect(optgroups).toHaveCount(2);
+      await expect(options).toHaveCount(2);
+      await expect(optgroups.nth(0)).toBeVisible();
+      await expect(options.nth(0)).toBeVisible();
+      await expect(optgroups.nth(1)).toBeVisible();
+      await expect(options.nth(1)).toBeVisible();
+    });
+
+    test('should reset filter value and show all optgroups and options again after closing and reopening again', async ({
+      page,
+    }) => {
+      await initSelect(page, {
+        props: {
+          name: 'Some name',
+          filter: true,
+        },
+        options: {
+          includeOptgroups: true,
+          values: [[{ value: '1a' }], [{ value: '2a' }]],
+        },
+      });
+      const host = getHost(page);
+      const buttonElement = getButton(page);
+      const filterElement = getFilter(page);
+      const filterInputElement = getFilterInput(page);
+      const options = getSelectOptions(page);
+      const optgroups = getSelectOptgroups(page);
+      const dropdown = getDropdown(page);
+
+      await buttonElement.click();
+
+      await expect(dropdown).toBeVisible();
+      await expect(filterElement).toBeFocused();
+      await filterInputElement.fill('2');
+
+      await expect(optgroups).toHaveCount(2);
+      await expect(options).toHaveCount(2);
+      await expect(optgroups.nth(0)).toBeHidden();
+      await expect(options.nth(0)).toBeHidden();
+      await expect(optgroups.nth(1)).toBeVisible();
+      await expect(options.nth(1)).toBeVisible();
+
+      await page.keyboard.press('Escape');
+      await expect(dropdown).toBeHidden();
+      await expect(host).toHaveJSProperty('value', undefined);
+      await expect(buttonElement).toBeFocused();
+
+      await buttonElement.click();
+
+      await expect(optgroups).toHaveCount(2);
+      await expect(options).toHaveCount(2);
+      await expect(optgroups.nth(0)).toBeVisible();
+      await expect(options.nth(0)).toBeVisible();
+      await expect(optgroups.nth(1)).toBeVisible();
+      await expect(options.nth(1)).toBeVisible();
+    });
+  });
+
+  test.describe('click', () => {
+    test('should not close dropdown if input is clicked', async ({ page }) => {
+      await initSelect(page, { props: { name: 'Some name', filter: true } });
+      const buttonElement = getButton(page);
+      const dropdown = getDropdown(page);
+      const filterElement = getFilter(page);
+
+      await buttonElement.click();
+      await expect(dropdown).toBeVisible();
+
+      await filterElement.click();
+      await expect(dropdown).toBeVisible();
+    });
+  });
+
+  test.describe('keyboard behavior', () => {
+    skipInBrowsers(['webkit']); // Safari focus management does not work correctly in Playwright
+
+    test('should focus filter input on Space key', async ({ page }) => {
+      await initSelect(page, { props: { name: 'Some name', filter: true } });
+      const host = getHost(page);
+      const filterElement = getFilter(page);
+
+      await page.keyboard.press('Tab');
+      await page.keyboard.press('Space');
+
+      await expect(filterElement).toBeFocused();
+      await expect(host).toHaveJSProperty('value', undefined);
+    });
+
+    test('should focus filter input on Enter key', async ({ page }) => {
+      await initSelect(page, { props: { name: 'Some name', filter: true } });
+      const host = getHost(page);
+      const filterElement = getFilter(page);
+
+      await page.keyboard.press('Tab');
+      await page.keyboard.press('Enter');
+
+      await expect(filterElement).toBeFocused();
+      await expect(host).toHaveJSProperty('value', undefined);
+    });
+
+    test('should focus filter input on ArrowDown key', async ({ page }) => {
+      await initSelect(page, { props: { name: 'Some name', filter: true } });
+      const host = getHost(page);
+      const filterElement = getFilter(page);
+
+      await page.keyboard.press('Tab');
+      await page.keyboard.press('ArrowDown');
+
+      await expect(filterElement).toBeFocused();
+      await expect(host).toHaveJSProperty('value', undefined);
+    });
+
+    test('should cycle through options while having filter input focused', async ({ page }) => {
+      await initSelect(page, { props: { name: 'Some name', filter: true } });
+      const host = getHost(page);
+      const filterElement = getFilter(page);
+
+      await page.keyboard.press('Tab');
+      await page.keyboard.press('Space');
+
+      await expect(filterElement).toBeFocused();
+      await filterElement.press('ArrowDown');
+
+      await expect(page.locator('p-select p-select-option').first()).toHaveJSProperty('highlighted', true);
+
+      await filterElement.press('Enter');
+
+      await expect(host).toHaveJSProperty('value', 'a');
+    });
+
+    test('should reset/keep highlighted option on filter input', async ({ page }) => {
+      await initSelect(page, { props: { name: 'Some name', filter: true } });
+      const filterElement = getFilter(page);
+      const filterInputElement = getFilterInput(page);
+      const options = getSelectOptions(page);
+
+      await page.keyboard.press('Tab');
+      await page.keyboard.press('Space');
+
+      await expect(filterElement).toBeFocused();
+
+      await expect(options.nth(0)).toHaveJSProperty('highlighted', undefined); // undefined since never was highlighted
+      await expect(options.nth(0)).toBeVisible();
+      await expect(options.nth(1)).toHaveJSProperty('highlighted', undefined); // undefined since never was highlighted
+      await expect(options.nth(1)).toBeVisible();
+      await expect(options.nth(2)).toHaveJSProperty('highlighted', undefined); // undefined since never was highlighted
+      await expect(options.nth(2)).toBeVisible();
+
+      await filterElement.press('ArrowDown');
+
+      await expect(options.nth(0)).toHaveJSProperty('highlighted', true);
+      await expect(options.nth(0)).toBeVisible();
+      await expect(options.nth(1)).toHaveJSProperty('highlighted', undefined); // undefined since never was highlighted
+      await expect(options.nth(1)).toBeVisible();
+      await expect(options.nth(2)).toHaveJSProperty('highlighted', undefined); // undefined since never was highlighted
+      await expect(options.nth(2)).toBeVisible();
+
+      await filterInputElement.fill('b');
+
+      // Highlight of option a will be reset since it doesn't match the filter
+      await expect(options.nth(0)).toHaveJSProperty('highlighted', false);
+      await expect(options.nth(0)).toBeHidden();
+      await expect(options.nth(1)).toHaveJSProperty('highlighted', undefined); // undefined since never was highlighted
+      await expect(options.nth(1)).toBeVisible();
+      await expect(options.nth(2)).toHaveJSProperty('highlighted', undefined); // undefined since never was highlighted
+      await expect(options.nth(2)).toBeHidden();
+
+      await filterElement.press('ArrowDown');
+
+      await expect(options.nth(0)).toHaveJSProperty('highlighted', false);
+      await expect(options.nth(0)).toBeHidden();
+      await expect(options.nth(1)).toHaveJSProperty('highlighted', true); // undefined since never was highlighted
+      await expect(options.nth(1)).toBeVisible();
+      await expect(options.nth(2)).toHaveJSProperty('highlighted', undefined); // undefined since never was highlighted
+      await expect(options.nth(2)).toBeHidden();
+
+      await filterInputElement.fill('');
+
+      await expect(options.nth(0)).toHaveJSProperty('highlighted', false);
+      await expect(options.nth(1)).toBeVisible();
+      await expect(options.nth(1)).toHaveJSProperty('highlighted', true);
+      await expect(options.nth(1)).toBeVisible();
+      await expect(options.nth(2)).toHaveJSProperty('highlighted', undefined); // undefined since never was highlighted
+      await expect(options.nth(1)).toBeVisible();
+
+      await filterElement.press('ArrowDown');
+
+      await expect(options.nth(0)).toHaveJSProperty('highlighted', false);
+      await expect(options.nth(1)).toHaveJSProperty('highlighted', false);
+      await expect(options.nth(2)).toHaveJSProperty('highlighted', true);
+
+      await filterInputElement.fill('c');
+
+      await expect(options.nth(0)).toHaveJSProperty('highlighted', false);
+      await expect(options.nth(0)).toBeHidden();
+      await expect(options.nth(1)).toHaveJSProperty('highlighted', false);
+      await expect(options.nth(1)).toBeHidden();
+      await expect(options.nth(2)).toHaveJSProperty('highlighted', true);
+      await expect(options.nth(2)).toBeVisible();
+    });
+
+    test('should not close select when Space character is typed into filter', async ({ page }) => {
+      await initSelect(page, {
+        props: { name: 'Some name', filter: true },
+        options: { values: [{ value: 'option a' }, { value: 'option b' }, { value: 'option c' }] },
+      });
+      const filterElement = getFilter(page);
+      const filterInputElement = getFilterInput(page);
+      const dropdown = getDropdown(page);
+      const options = getSelectOptions(page);
+
+      await page.keyboard.press('Tab');
+      await page.keyboard.press('Space');
+
+      await expect(dropdown).toBeVisible();
+      await expect(filterElement).toBeFocused();
+      await filterElement.pressSequentially('option');
+      await filterInputElement.press('Space');
+      await filterInputElement.press('a');
+
+      // Dropdown should stay open
+      await expect(dropdown).toBeVisible();
+      await filterElement.press('ArrowDown');
+
+      await expect(options.nth(0)).toHaveJSProperty('highlighted', true);
+      await expect(options.nth(0)).toBeVisible();
+      await expect(options.nth(1)).toBeHidden();
+      await expect(options.nth(2)).toBeHidden();
+    });
+  });
+});
+
 test.describe('click events', () => {
   test('should open dropdown on mouseclick and close dropdown on 2nd click', async ({ page }) => {
     await initSelect(page);
@@ -1319,10 +1841,40 @@ test.describe('lifecycle', () => {
     await option1.click();
     await waitForStencilLifecycle(page);
 
-    const status2 = await getLifecycleStatus(page);
-    expect(status2.componentDidUpdate['p-select-option'], 'componentDidUpdate: p-select-option').toBe(2);
-    expect(status2.componentDidUpdate['p-select'], 'componentDidUpdate: p-select').toBe(2);
-    expect(status2.componentDidUpdate.all, 'componentDidUpdate: all').toBe(4);
+    // Use polling to fix flakiness
+    await expect
+      .poll(
+        async () => {
+          const status = await getLifecycleStatus(page);
+          return status.componentDidUpdate['p-select-option'];
+        },
+        {
+          message: 'componentDidUpdate: p-select-option',
+        }
+      )
+      .toBe(2);
+    await expect
+      .poll(
+        async () => {
+          const status = await getLifecycleStatus(page);
+          return status.componentDidUpdate['p-select'];
+        },
+        {
+          message: 'componentDidUpdate: p-select',
+        }
+      )
+      .toBe(2);
+    await expect
+      .poll(
+        async () => {
+          const status = await getLifecycleStatus(page);
+          return status.componentDidUpdate.all;
+        },
+        {
+          message: 'componentDidUpdate: all',
+        }
+      )
+      .toBe(4);
   });
 
   skipInBrowsers(['webkit'], () => {