+++ packages/components/src/components/textarea/textarea-utils.ts
@@ -1,7 +1,7 @@
 import type { FormState } from '../../utils/form/form-state';
 export type TextareaState = FormState;
 
-export const AUTO_COMPLETE = ['off', 'on', ''] as const;
+export const AUTO_COMPLETE = ['off', 'on'] as const;
 export type TextareaAutoComplete = (typeof AUTO_COMPLETE)[number];
 
 export const TEXTAREA_WRAPS = ['hard', 'soft', 'off'] as const;